Add files via upload
#Objective: 
The objective of the jupyter notebook is to build good model to predict how much area would be burn during fire.
The data can be used for both classification and regression task. Several data scientist converted the target variable which is 
continous in nature into binary categorial variable to perform the classification task in order to predict whether there would be burnt or not. However, our approach have been first to visualize the data and later perform regression task.

#Details of the dataset

   This is a very difficult regression task. It can be used to test regression methods. Also,
   it could be used to test outlier detection methods, since it is not clear how many outliers
   are there. Yet, the number of examples of fires with a large burned area is very small.

5. Number of Instances: 517 

6. Number of Attributes: 12 + output attribute
  
   Note: several of the attributes may be correlated, thus it makes sense to apply some sort of
   feature selection.

7. Attribute information:

   For more information, read [Cortez and Morais, 2007].

   1. X - x-axis spatial coordinate within the Montesinho park map: 1 to 9
   2. Y - y-axis spatial coordinate within the Montesinho park map: 2 to 9
   3. month - month of the year: "jan" to "dec" 
   4. day - day of the week: "mon" to "sun"
   5. FFMC - FFMC index from the FWI system: 18.7 to 96.20
   6. DMC - DMC index from the FWI system: 1.1 to 291.3 
   7. DC - DC index from the FWI system: 7.9 to 860.6 
   8. ISI - ISI index from the FWI system: 0.0 to 56.10
   9. temp - temperature in Celsius degrees: 2.2 to 33.30
   10. RH - relative humidity in %: 15.0 to 100
   11. wind - wind speed in km/h: 0.40 to 9.40 
   12. rain - outside rain in mm/m2 : 0.0 to 6.4 
   13. area - the burned area of the forest (in ha): 0.00 to 1090.84 
   (this output variable is very skewed towards 0.0, thus it may make
    sense to model with the logarithm transform). 

8. Missing Attribute Values: None


# Conclusion
The purpose of the project was to predict the value of burnt area
Which is impacted by various features like month, ISI index, temperature
Rain etc.

There was some duplicated values in the data set that has been removed
The null values was not present in the data

While performing hypothesis test like spearman to find the significance of
Correlation between numerical independent and response variable it was observed
There was very weak correlation between independent and response feature (area)
.It was improved using log(x+1) transformation

ANOVA test shows that month is significant variables that affects the area while days
Do not. Post hoc analysis accomplished with Dunn’s test conveys that pair of month
Like dec and oct, dec and march, and aug and dec impacts the area significantly.

Severals models were built and evaluated using mean squared error.OLS methods
Depicts that with log(x+1) transformation we can explain the variance of 40% with a simple model. This is the simple model where wald test confirms only four feature were significant
To predict burnt area, the four feature namely are FFM, ISI index , wind and temperature 


Later we tuned the Gradient Boosted Regressor where features were selected using
Recursive feature elimination.  The best performance was obtained with max depth =2
And minimum split = 14

Bayesian regressor was attempted on the model. The performance was good in test data
Not on training data which indicate the presence of high bias in the model. 

Finally we can select OLS best logistic regression model for predictive modelling as it is light
And performing similar to ensemble model.  

All model are showing high biases, this is because the response variables contains almost
50% of the data with zero area, if somehow more  records are taken containing different values other than zero, the model will perform good.